Added Authorization

+++ router/Posts.py
@@ -9,24 +9,20 @@ import models
 from database import get_db
 from schemas import PostResponse,Postupdate,Postcreate
 
+from auth import Currentuser
+
 
 router=APIRouter()
 
 
 @router.post("",response_model=PostResponse,status_code=status.HTTP_201_CREATED)
-async def create_post(post:Postcreate,db:Annotated[AsyncSession,Depends(get_db)]):
+async def create_post(post:Postcreate,current_user:Currentuser,db:Annotated[AsyncSession,Depends(get_db)]):
 
-    result=await db.execute(select(models.User).where(models.User.id==post.user_id))
-    user=result.scalars().all()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found"
-        )
+
     new_post=models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
           )
     
     db.add(new_post)
@@ -56,6 +52,7 @@ async def get_post(post_id :int,db:Annotated[AsyncSession,Depends(get_db)]):
 @router.put("/{post_id}",response_model=PostResponse)
 async def Update_post(post_id :int,
              post_data:Postcreate,
+             current_user:Currentuser,
              db:Annotated[AsyncSession,Depends(get_db)]):
     result= await db.execute(select(models.Post)
                              .options(selectinload(models.Post.author))
@@ -66,20 +63,14 @@ async def Update_post(post_id :int,
         status_code=status.HTTP_404_NOT_FOUND,
         detail="Posts not found")
     
-    if post_data.user_id != post.user_id:
-        result=await db.execute(select(models.User).where(models.User.id==post.user_id))
-     
-        user=result.scalars().first()
-
-        if not user:
-            raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User posts not found"
+    if post.user_id!=current_user.id:
+        raise HTTPException(
+              status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post ",
         )
 
     post.title=post_data.title
     post.content=post_data.content
-    post.user_id=post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -89,6 +80,7 @@ async def Update_post(post_id :int,
 @router.patch("/{post_id}",response_model=PostResponse)
 async def Update_post_partial(post_id :int,
              post_data:Postupdate,
+             current_user:Currentuser,
              db:Annotated[AsyncSession,Depends(get_db)]):
     result=await db.execute(select(models.Post).where(models.Post.id==post_id))
     post=result.scalars().first()
@@ -97,7 +89,12 @@ async def Update_post_partial(post_id :int,
         status_code=status.HTTP_404_NOT_FOUND,
         detail="Posts not found")
     
-
+    if post.user_id!=current_user.id:
+        raise HTTPException(
+              status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post ",
+        )
+    
     Update_post=post_data.model_dump(exclude_unset=True)
     for field,value in Update_post.items():
         setattr(post,field,value)
@@ -107,7 +104,7 @@ async def Update_post_partial(post_id :int,
     return post
 
 @router.delete("/{post_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id :int,db:Annotated[AsyncSession,Depends(get_db)]):
+async def delete_post(post_id :int,current_user:Currentuser,db:Annotated[AsyncSession,Depends(get_db)]):
     result=await db.execute(select(models.Post).where(models.Post.id==post_id))
     post=result.scalars().first()
     if not post:
@@ -115,6 +112,11 @@ async def delete_post(post_id :int,db:Annotated[AsyncSession,Depends(get_db)]):
         status_code=status.HTTP_404_NOT_FOUND,
         detail="Posts not found"
     )
+    if post.user_id!=current_user.id:
+        raise HTTPException(
+              status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post ",
+        )
 
     await db.delete(post)
     await db.commit()
